+++ packages/web/src/hooks/index.ts
@@ -111,5 +111,5 @@ export {
 } from './useChapters';
 
 // Intake
-export { useDecompose, useDetectDuplicates, useCommitEntities } from './useIntake';
+export { useDecompose, useCommitEntities } from './useIntake';
 export { useIntakeImport } from './useIntakeImport';